requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cssselect==1.2.0
fake-useragent==1.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
parse==1.19.1
pathlib==1.0.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.6
zipp==3.16.2
